+++ CHANGELOG.md
@@ -7,3 +7,15 @@
 - Parse command arguments
 - Scan Postgres
 - Clear destination folder
+- Create data project
+- Create SQL script
+  - Columns
+  - Primary keys
+  - Foreign keys
+  - Indexes
+- Create entities
+  - Data annotations applied
+    - Key, Required, MaxLength, DisplayName
+- Create EF contexts
+  - InMemory
+  - Postgres